our project is large enough now... moved models into separate folder with one class per file
