+++ invadedhouse/__init__.py
@@ -1,8 +1,5 @@
 from .commands import CommandInterpreter
-from .player import Player, Inventory
-from .rooms import Room, Monster, Item, Weapon, House
 
 __all__ = [
-    'CommandInterpreter', 'House', 'Inventory', 'Item', 'Monster',
-    'Player', 'Room', 'Weapon'
+    'CommandInterpreter'
 ]
