+++ invadedhouse/models/__init__.py
@@ -1,0 +1,13 @@
+from .chest import Chest
+from .house import House
+from .inventory import Inventory
+from .item import Item
+from .monster import Monster
+from .player import Player
+from .room import Room
+from .weapon import Weapon
+
+__all__ = [
+    'Chest', 'House', 'Inventory', 'Item',
+    'Monster', 'Player', 'Room', 'Weapon'
+]
